% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.r
\name{scale_fill_weird}
\alias{scale_fill_weird}
\title{Fill scale constructor for weird colors}
\usage{
scale_fill_weird(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in weird_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Fill scale constructor for weird colors
}
