% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.r
\name{weird_cols}
\alias{weird_cols}
\title{Extract Weird Data Science colours as hex codes}
\usage{
weird_cols(...)
}
\arguments{
\item{...}{Character names of weird_palette colours.}
}
\description{
Extract Weird Data Science colours as hex codes
}
