% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.r
\name{weird_pal}
\alias{weird_pal}
\title{Interpolate a weird data science color palette}
\usage{
weird_pal(palette = "weird", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Interpolate a weird data science color palette
}
